backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
aiohttp==3.7.4
nltk==3.6.3
MetaTrader5==5.0.37
python-dotenv==0.19.0